[Move] APIRegisterer, UnsafeRunnable
