+++ plugin/src/main/java/net/okocraft/box/plugin/util/UnsafeRunnable.java
@@ -1,4 +1,4 @@
-package net.okocraft.box.util;
+package net.okocraft.box.plugin.util;
 
 import java.util.concurrent.CompletionException;
